Feat: 상품 조회 api router 상품 데이터에 유저 데이터 및 신고여부 포함 - 상품 조회시 유저 데이터와 상품 정보를 한번에 받아오도록 쿼리문 추가 - 신고여부를 반환하기 위해 isReport를 상품 데이터에 추가하여 반환

+++ pages/api/product/[productId]/index.ts
@@ -1,13 +1,16 @@
 import { ERROR_MESSAGE } from "@/constants/constant";
 import mongoose from "mongoose";
 
-import { Product } from "@/lib/db/schema";
-import { checkAuthorization } from "@/lib/server";
+import { Product, User } from "@/lib/db/schema";
+import { checkAuthorization, sessionOptions } from "@/lib/server";
 import { NextApiRequest, NextApiResponse } from "next";
 import dbConnect from "@/lib/db";
 import { deleteObject, ref } from "firebase/storage";
 import { storage } from "@/lib/firebaseSetting";
-import { ProductImgData } from '@/types/productTypes';
+import { ProductImgData } from "@/types/productTypes";
+import { getIronSession } from "iron-session";
+import { IronSessionType } from "@/types/apiTypes";
+import { verifyToken } from "@/lib/token";
 
 export default async function handler(
   req: NextApiRequest,
@@ -15,6 +18,21 @@ export default async function handler(
 ) {
   if (req.method === "GET") {
     try {
+      const session = await getIronSession<IronSessionType>(
+        req,
+        res,
+        sessionOptions
+      );
+
+      const accessToken = session.accessToken;
+
+      const decodeAccessToken = verifyToken(
+        accessToken as string,
+        process.env.NEXT_SECRET_ACCESS_TOKEN_KEY as string
+      );
+
+      const uid = decodeAccessToken?.data?.user.uid;
+
       const { productId } = req.query;
 
       if (!productId) {
@@ -40,11 +58,131 @@ export default async function handler(
         return;
       }
 
-      res.status(200).json({ message: "상품 조회에 성공했어요.", product });
+      const isReport = product.reportUserIds.includes(uid);
+
+      // 유저 프로필, 리뷰점수 및 최신 상품 목록을 조인합니다.
+      const aggregation = [
+        {
+          $match: { _id: new mongoose.Types.ObjectId(product.uid as string) },
+        },
+        {
+          $lookup: {
+            from: "reviewScore",
+            localField: "uid",
+            foreignField: "uid",
+            as: "reviewInfo",
+          },
+        },
+        {
+          $unwind: {
+            path: "$reviewInfo",
+            preserveNullAndEmptyArrays: true,
+          },
+        },
+        {
+          $addFields: {
+            isFollower: {
+              $cond: {
+                if: { $eq: [uid, undefined] }, 
+                then: false,
+                else: {
+                  $in: [
+                    new mongoose.Types.ObjectId(uid as string),
+                    "$followers",
+                  ],
+                },
+              },
+            },
+          },
+        },
+        {
+          $addFields: {
+            reviewPercentage: {
+              $cond: {
+                if: {
+                  $eq: [{ $ifNull: ["$reviewInfo.totalScore", null] }, null],
+                },
+                then: 0,
+                else: {
+                  $round: [
+                    {
+                      $multiply: [
+                        {
+                          $divide: [
+                            "$reviewInfo.totalScore",
+                            "$reviewInfo.totalReviewCount",
+                          ],
+                        },
+                        20,
+                      ],
+                    },
+                    1,
+                  ],
+                },
+              },
+            },
+          },
+        },
+        {
+          $addFields: {
+            convertedProductIds: {
+              $map: {
+                input: "$productIds",
+                as: "id",
+                in: { $toObjectId: "$$id" },
+              },
+            },
+          },
+        },
+        {
+          $lookup: {
+            from: "products",
+            localField: "convertedProductIds",
+            foreignField: "_id",
+            as: "recentProducts",
+          },
+        },
+        {
+          $addFields: {
+            recentProducts: {
+              $filter: {
+                input: "$recentProducts",
+                as: "product",
+                cond: { $ne: ["$$product.block", true] },
+              },
+            },
+          },
+        },
+        {
+          $project: {
+            _id: 1,
+            email: 1,
+            nickname: 1,
+            profileImg: 1,
+            isFollower: 1,
+            followers: 1,
+            followings: 1,
+            productIds: 1,
+            reviewPercentage: 1,
+            recentProducts: 1,
+          },
+        },
+      ];
+
+      const userWithReviews = await User.aggregate(aggregation);
+
+      res.status(200).json({
+        message: "상품 조회에 성공했어요.",
+        product: {
+          ...product._doc,
+          isReport,
+          auth: { ...userWithReviews[0], uid: userWithReviews[0]._id },
+        },
+      });
     } catch (error) {
-      console.error("상품 조회 에러", error);
+      console.error(error);
       res.status(500).json({
-        message: ERROR_MESSAGE,
+        message: "상품 조회에 실패했어요\n잠시 후 다시 시도해주세요.",
       });
     }
   }
@@ -139,7 +277,9 @@ export default async function handler(
           .json({ message: "잘못된 요청이에요. 로그인 정보를 확인해주세요." });
         return;
       }
-      const productImgNameArray = product.imgData.map((data: ProductImgData) => data.name);
+      const productImgNameArray = product.imgData.map(
+        (data: ProductImgData) => data.name
+      );
 
       try {
         const removeImgPromise = productImgNameArray.map((name: string) => {
@@ -150,6 +290,24 @@ export default async function handler(
       } catch (error) {
         console.error(error);
         console.log("상품 이미지 삭제에 실패했어요.");
+        res.status(500).json({
+          message: "상품 삭제에 실패했어요.\n잠시 후 다시 시도해주세요.",
+        });
+        return;
+      }
+
+      const profileResult = await User.updateOne({
+        $pull: {
+          productIds: new mongoose.Types.ObjectId(product._id as string),
+        },
+      });
+
+      if (!profileResult.acknowledged || profileResult.modifiedCount === 0) {
+        console.error("유저 프로필에서 상품 아이디 삭제에 실패했어요.");
+        res.status(500).json({
+          message: "상품 삭제에 실패했어요.\n잠시 후 다시 시도해주세요.",
+        });
+        return;
       }
 
       const result = await Product.deleteOne({
@@ -157,6 +315,7 @@ export default async function handler(
       });
 
       if (!result.acknowledged || result.deletedCount === 0) {
+        console.error("상품 삭제에 실패했어요.");
         res.status(500).json({
           message: "상품 삭제에 실패했어요.\n잠시 후 다시 시도해주세요.",
         });
